pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wsr"
version = "0.1.0"
description = "Wayland Session Recorder (Python Port of xsr)"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "evdev",
    "Pillow",
]

[project.scripts]
wsr = "wsr.main:main"
